003-ai-agent/003-001-first-ai-agent/pyproject.toml
[project]
name = "003-001-first-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.15.0",
    "openai-agents>=0.6.5",
    "python-dotenv>=1.2.1",
    "xai-sdk>=1.5.0",
    "youtube-transcript-api>=1.2.3",
]

[dependency-groups]
dev = ["watchfiles>=1.0.0"]
